move ytdl module into lib/ dir & organize imports
